Start branch removal. Optimize perft function.

+++ src/test/kotlin/in/kyle/chess/bitboard/TestBitboard.kt
@@ -6,7 +6,7 @@ import io.kotest.matchers.ints.shouldBeExactly
 
 class TestBitboard : StringSpec({
     "bitboard scans should contain next index" {
-        val testValue = 0b1000UL
+        val testValue = 0b1000UL.toLong()
         val testScan = bitscanForward(testValue)
         testScan shouldBeExactly 3
     }
